BackEnd - fixed previous commit

+++ src/com/maci/mapper/StudentMapper.java
@@ -12,7 +12,7 @@ public class StudentMapper implements RowMapper<Student>{
 	@Override
 	public Student mapRow(ResultSet rs, int arg1) throws SQLException {
 		Student student=new Student();
-		student.setId(rs.getLong("id"));
+		student.setId(rs.getInt("id"));
 		student.setIme(rs.getString("ime"));
 		student.setPrezime(rs.getString("prezime"));
 		return student;
